refactor(client): move fulfillment slice into features/fulfillment
