+++ client/src/features/fulfillment/components/delivery/DeliveryTimelineDialog.tsx
@@ -5,10 +5,10 @@ import {
   DialogHeader,
   DialogTitle,
   DialogDescription,
-} from '../../shared/ui/dialog';
-import StatusBadge from '../../shared/components/StatusBadge';
-import { formatDateTime } from '../../shared/lib/utils';
-import { useTranslation } from '../../shared/i18n/index';
+} from '../../../../shared/ui/dialog';
+import StatusBadge from '../../../../shared/components/StatusBadge';
+import { formatDateTime } from '../../../../shared/lib/utils';
+import { useTranslation } from '../../../../shared/i18n/index';
 
 import type { DeliveryStatusHistoryEntry } from '@/types';
