+++ client/src/features/fulfillment/pages/Deliveries.test.tsx
@@ -2,10 +2,10 @@ import { describe, it, expect, beforeEach } from 'vitest';
 import type { ReactNode } from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { TransportProvider } from '../shared/lib/transport/index';
-import { createMemoryTransport, type MemoryTransport } from '../shared/lib/transport/memory';
-import { useSettingsStore } from '../shared/store/settingsStore';
-import { useAuthStore } from '../features/auth/store/authStore';
+import { TransportProvider } from '../../../shared/lib/transport/index';
+import { createMemoryTransport, type MemoryTransport } from '../../../shared/lib/transport/memory';
+import { useSettingsStore } from '../../../shared/store/settingsStore';
+import { useAuthStore } from '../../auth/store/authStore';
 import type { DeliveryOrder, DeliveryPerformance } from '@/types';
 import Deliveries from './Deliveries';
