+++ client/src/features/fulfillment/pages/OnlineOrders.tsx
@@ -1,17 +1,17 @@
 import { useState } from 'react';
 import { Eye, Truck, XCircle, CheckCircle, Package } from 'lucide-react';
-import { formatCurrency } from '../shared/lib/utils';
-import { Button } from '../shared/ui/button';
-import { Badge } from '../shared/ui/badge';
+import { formatCurrency } from '../../../shared/lib/utils';
+import { Button } from '../../../shared/ui/button';
+import { Badge } from '../../../shared/ui/badge';
 import {
   Dialog,
   DialogContent,
   DialogHeader,
   DialogTitle,
   DialogDescription,
-} from '../shared/ui/dialog';
-import { useTranslation } from '../shared/i18n/index';
-import { resource } from '../shared/lib/resource';
+} from '../../../shared/ui/dialog';
+import { useTranslation } from '../../../shared/i18n/index';
+import { resource } from '../../../shared/lib/resource';
 import type { OnlineOrder } from '@/types';
 
 const onlineOrders = resource<OnlineOrder>('online-orders');
